+++ script/circle/bintray-deploy.sh
@@ -1,7 +1,5 @@
 #!/bin/bash
 
-set -x
-
 curl -f -u$BINTRAY_USERNAME:$BINTRAY_API_KEY -X GET \
   https://api.bintray.com/repos/docker-compose/${CIRCLE_BRANCH}